GTD Data Analysis in Python
Designed a Python project that parses an Excel sheet of the Global Terrorism Database and analyses data to provide the user with useful information. Done under the guidance of Dr. Charles Severance from the University of Michigan as part of the Python for Everybody Coursera specialization. Accessing this resource is authorized only for editing purposes by the owner and showing the project to potential employees and clients. Accessing this course to cheat in the Coursera specialization violates the Coursera code of conduct. The owner claims no responsibility for usage of this code for such purposes.